update testcase of partition
